preinstall and postuninstall support
This patch adds support for two special daemonsets
 - a pre-install daemonset which is run before the actual installation
 - a post-uninstall daemonset which is run after the actual
   uninstallation

The CRD gets new fields to define environment variables, volumes and
volume sets for the daemon sets and the sample CR yaml is extended with
working examples. There are many new fields added here, and it may seem
a bit much. The reasoning is that we want to keep the two different
daemon sets configurable individually as there can be different
requirements for both. To lessen the clutter in the CRD we create
sub-structures for pre-install and post-uninstall configurations.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>