scorecard tests: disable some tests
I want to disable specDescriptor and statusDescriptor tests because
we don't use both at the moment. We can add those in the future but it's
a cosmetic add-on that we can enable for the next release. The impact is
that CRD spec fields that users can fill in via Operatorhub gui based
installation would look nicer.

Let's also disable the test that checks that all used resources by a CDR
are specified in the CSV. We don't specify the resources at the moment.
It's also something we can change for the next release but it's not
critical.

Signed-off-by: Jens Freimann <jfreimann@redhat.com>